Added StatsServiceTest

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -1,7 +1,7 @@
 package ru.netology.stats;
 
 public class StatsService {
-    public int minSales(long[] sales) {
+    public int minSales(int[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -12,5 +12,6 @@ public class StatsService {
 
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
+
 
 }
